backend/requirements.txt
asgiref==3.10.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.29.0
pillow==12.0.0
sqlparse==0.5.5
tzdata==2025.3
